Move getLocalAddresses to NetUtil
